Ability to push all parking info into 1 document
I feel pain and suffering but there you go.
The structure I followed is that you have and array of hashmaps. these hashmaps are hashmaps of another hashmaps such that you have the car park id as the key and the value is an array of the rest of the data as hashmaps. I feel pain. Im sorry